Prevent canonical exponent accumulation from overflowing
Retain multiplication when positive exponent sums exceed the supported int range, and order multivariate terms by a long total degree. Add regressions for overflow, the exact boundary, and both degree-sorting paths.

Two regression methods fail against the unchanged main sources; all 575 core tests pass with the fix under Maven and Java 25. No quality gates, baselines, dependencies, or archived evidence changed.

Refs #661.

+++ regelsuche-core/src/main/java/de/regelsuche/canonical/PolynomialNormalizer.java
@@ -184,8 +184,8 @@ public final class PolynomialNormalizer {
             return new Monomial(result);
         }
 
-        private int degree() {
-            int degree = 0;
+        private long degree() {
+            long degree = 0;
             for (int exponent : powers.values()) {
                 degree += exponent;
             }
@@ -233,7 +233,7 @@ public final class PolynomialNormalizer {
         private static final Comparator<
             Map.Entry<Monomial, ExactRational>> TERM_ORDER =
                 Comparator
-                    .<Map.Entry<Monomial, ExactRational>>comparingInt(
+                    .<Map.Entry<Monomial, ExactRational>>comparingLong(
                         entry -> entry.getKey().degree())
                     .reversed()
                     .thenComparing(
